+++ src/Morrigu/Core/MouseCodes.h
@@ -9,7 +9,7 @@
 
 namespace MRG
 {
-	using MouseCode = uint16_t;
+	using MouseCode = int;
 	namespace Mouse
 	{
 		enum : MouseCode
